Utilise les meme filtres entre la carte et la liste pour harmoniser
